Update instructor-notes.md
Adding a tip of a common error, or where a purposeful error can be introduced in lesson 2 "starting with data". Trying to get data from a directory the student is not in, or setting the destination for a file to somewhere that does not exist is a common issue I've seen when working with students new to R and coding in general. I believe this is a nice simple error to introduce and walk through breaking down an error code, as well as a good spot to briefly introduce navigating directories.